Don't search for matching word stems
This may be something we want to use in future, but Elasticlunr doesn't
expose the stem that is being searched for, e.g. searching for
"testing" would return "test", but wouldn't expose the stem that had
returned the match, meaning we can't highlight the matching word in the
search result excerpt.

[This StackOverflow
answer](https://stackoverflow.com/questions/28346291/lunr-return-the-stem-of-the-searched-terms-so-i-can-highlight-it-within-results)
suggests a way of possibly doing this, but I don't want to introduce too
much clientside Javascript looping behaviour here as it might slow
things down.