Clean up newlines and spaces in search results
This modifies the Jekyll/Liquid code to send the content through to the window
with newlines still in tact. Without any further changes, they would be
displayed as spaces (with consecutive spaces collapsed down to one)

Then in the JavaScript search script, a new formatContent method cleans up the
new lines and spaces, by:
- replacing common sentence-final punctuation (.?!) followed by two or more new
  lines or spaces with the punctuation followed by one space
- replacing any remaining instances of two or more new lines with a full stop
  followed by a space
- replacing any remaining new lines (including individual instances) with a space

A few of the Markdown files have formatting that differs to others, e.g. line
breaks within paragraphs to keep each line of code short, or indentation in some
cases where paragraphed lists are used. The current regex and Lunr deal with
most if not all of this, but we may want to reduce the chaos with some Markdown
linting in future