Make search input responsive and use BEM class
This creates an extra search input below the navbar `ul`. This will be
displayed in block format on smaller screens, while the inline input
will be displayed on larger screens. The script is also amended to
prepopulate both of these inputs when a query URL param is present

We're adding the BEM class here rather than a universal
`input[type="text"]` selector in the typography stylesheet. The
typography stylesheet is doing more than typography still, so we might
want to address that in the future