Remove search page from search results
In Jekyll's `_site` folder, there was a `search.html` page with all the
`window.store` contents added to it as a result of the script that runs on
loading the page. The script was also searching through the content it had
created on that page when looking for matches, leading to it showing up as a
strange-looking result. This commit prevents it from being added to
`window.store` and therefore filters it out of search results