Format search `innerHtml` code addition like HTML
Previously the code that added to the search `innerHtml` string was structured
so that any static code was added via a string on a single line until there was
a need for dynamic code, at which point it would be separated onto a new line.
This shifts the approach to make it read more like a typical HTML document,
using interpolated strings, line breaks and indentation. The result gives a
clearer picture of the structure without having to parse it as one continuous
string